CREATE blog landing page
I have now designed how the landing page should look, and I am now in the process of fleshing out this design, however first I wanted to create an element from the design which used a stretched text which had two colours, clipped horizontally. And this wasn't so hard to make, and then I wanted to turn it into a web component, because I have heard so much about that, and then I just implemented it, and it is really rather cool, because it is now easy to set the colours and height of the text and the clipover point.

Soon I will get the other things in place.